Model gracza i animacje i pare poprawek
Model gracza i animacje 
Naprawa player score - gdy miales full hp, power up nie dodawał score (null był)
Zmiana UI pause menu i restart menu tak żeby był widoczny score i distance
Zmiana czcionki w obydwu menu tak zeby pasowały do counterów z gry
