+++ Assets/Animation.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ba6658faf0e73c74cb05b243b14de058
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
