+++ Assets/Animation/PlayerMoves.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e802b71038c57ff43abaeac069e437fe
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
