--- Assets/PlayerMove.cs
@@ -1,46 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class PlayerMove : MonoBehaviour
-{
-    [SerializeField] private float moveSpeed = 7f;
-
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        transform.position += transform.forward * moveSpeed * Time.deltaTime;
-
-        Vector2 inputVector = new Vector2(0, 0);
-
-        if (Input.GetKey(KeyCode.W))
-        {
-            inputVector.y = +1;
-        }
-        if (Input.GetKey(KeyCode.S))
-        {
-            inputVector.y = -1;
-        }
-        if (Input.GetKey(KeyCode.A))
-        {
-            inputVector.x = -1;
-        }
-        if (Input.GetKey(KeyCode.D))
-        {
-            inputVector.x = +1;
-        }
-
-
-        Vector3 moveDir = new Vector3(inputVector.x, 0f, inputVector.y);
-
-        transform.position += moveDir * moveSpeed * Time.deltaTime;
-    }
-
-
-}
