--- Assets/PlayerMove.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 56dd9365f3d79d54b88dda800af02842
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
