+++ Assets/Scripts/Invulnerability.cs
@@ -8,6 +8,7 @@ public class Invulnerability : MonoBehaviour
     
     [SerializeField] private float invuDuration = 2f;
     
+    
 
 
     void Start()
@@ -26,7 +27,8 @@ public class Invulnerability : MonoBehaviour
 
     private void PlayerFlash()
     {
-        MeshRenderer meshRenderer = GetComponent<MeshRenderer>();
+
+        SkinnedMeshRenderer meshRenderer = GetComponent<SkinnedMeshRenderer>();
         Color currentColor = meshRenderer.material.color;
         currentColor.a = .2f;
         meshRenderer.material.DOColor(currentColor, .2f).SetLoops(10, LoopType.Yoyo);
