+++ Assets/Scripts/PlayerAnimator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 44c16ea1233a3b242866a177396d379c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
