+++ Assets/Tex.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1a856e8433b4b2047a0aa3366d57a1f7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
