+++ Assets/Tex/eyes.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4fd916762134eb140b883eb6dc409673
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
