+++ Assets/Tex/face.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 01288d6ccde398a449856a941e619c26
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
