+++ Assets/Tex/hat_glove_pants_shoes_belt_txt.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5b3cf40fb234aa44fa84f6bdf4b6b603
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
