+++ Assets/Tex/shirt_txt.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2e3521727e7da7e42a0596d11f729799
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
