+++ Assets/Tex/shotgun_txt.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 617e8c30d6f5fc848961ee49148a71c3
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
