+++ Assets/UI/TiltPrism-Regular SDF.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e636a2450b3fc1d488a7189d31d5ecb8
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
